requirements.txt
transformers
datasets
faiss-cpu
pillow
datasets
pandas
torch